Only revert tag if latest tag conflicts
The Anka Registry does not support deleting any random tag. Our only
options are to delete ALL tags or "revert" the latest. During a force
build, we only need to revert the latest tag if that tag matches the
tag we're trying to push. Otherwise, we can leave it alone.

This does prevent us from successfully re-pushing an older tag.

Signed-off-by: Tom Duffield <github@tomduffield.com>